rag-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.6
qdrant-client==1.8.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
asyncpg==0.29.0